requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
Werkzeug==2.3.7
dotenv
gunicorn
